Refactored handler functions into separate file
