+++ app/server.js
@@ -1,12 +1,8 @@
 const hapi = require('hapi');
 const pkg = require('./package');
-const es = require('elasticsearch');
-const R = require('ramda');
-const Boom = require('boom');
-// const eb = require('elastic-builder');
+const handlers = require('./lib/handlers/handlers');
 
 const server = new hapi.Server();
-const esClient = new es.Client({ host: 'http://localhost:9200' });
 
 server.connection({
   host: 'localhost',
@@ -28,78 +24,31 @@ server.route({
 server.route({
   method: 'GET',
   path: '/task',
-  handler: (request, reply) => // eslint-disable-line no-unused-vars
-    esClient.search({ index: 'tasks-out', type: 'task', size: 100 })
-      .then(result => result.hits.hits)
-      .then(R.map(R.prop('_source')))
-      .then(reply)
-      .catch(err => {
-        const statusCode = R.propOr(500, 'statusCode', err);
-        request.log(['error'], err.message);
-        return reply(Boom.wrap(err, statusCode));
-      })
+  handler: handlers.listHandler
 });
-
-const getHandler = (request, reply) => // eslint-disable-line no-unused-vars
-  esClient.get({ index: 'tasks-out', type: 'task', id: request.params.id })
-    .then(R.prop('_source'))
-    .then(reply)
-    .catch(err => {
-      const statusCode = R.propOr(500, 'statusCode', err);
-      request.log(['error'], err.message);
-      return reply(Boom.wrap(err, statusCode));
-    });
 
 server.route({
   method: 'GET',
   path: '/task/{id}',
-  handler: getHandler
+  handler: handlers.getHandler
 });
 
 server.route({
   method: 'POST',
   path: '/task',
-  handler: (request, reply) => { // eslint-disable-line no-unused-vars
-    return esClient.index({ index: 'tasks-in', type: 'task', body: request.payload })
-      .then(response => esClient.get({ index: 'tasks-out', type: 'task', id: response._id }))
-      .then(R.prop('_source'))
-      .then(reply)
-      .catch(err => {
-        const statusCode = R.propOr(500, 'statusCode', err);
-        request.log(['error'], err.message);
-        return reply(Boom.wrap(err, statusCode));
-      });
-  }
+  handler: handlers.postHandler
 });
 
 server.route({
   method: 'PUT',
   path: '/task/{id}',
-  handler: (request, reply) => {
-    return esClient.update({ index: 'tasks-in', type: 'task', id: request.params.id, body: { doc: request.payload } })
-      .then(response => esClient.get({ index: 'tasks-out', type: 'task', id: response._id }))
-      .then(R.prop('_source'))
-      .then(reply)
-      .catch(err => {
-        const statusCode = R.propOr(500, 'statusCode', err);
-        request.log(['error'], err.message);
-        return reply(Boom.wrap(err, statusCode));
-      });
-  }
+  handler: handlers.putHandler
 });
 
 server.route({
   method: 'DELETE',
   path: '/task/{id}',
-  handler: (request, reply) => {
-    return esClient.delete({ index: 'tasks-in', type: 'task', id: request.params.id })
-      .then(reply)
-      .catch(err => {
-        const statusCode = R.propOr(500, 'statusCode', err);
-        request.log(['error'], err.message);
-        return reply(Boom.wrap(err, statusCode));
-      });
-  }
+  handler: handlers.deleteHandler
 });
 
 if (!module.parent) {
